Update Messages
This update includes correcting the parse method, option to pass header instance, and header parsing.

This is an important addition, because it includes methods that ensure my header, response_id, sequence_id, & device_id is correct, and updated properly with incoming messages as well. This will allow for proper exception catches and throws throughout the unity file.