update assembly_extentions with mesh .obj export
Update Assembly Extensions with Mesh Assembly export. The difficult part here is there is multiple ways to make a mesh assembly, both with a frame and without a frame. I am not sure if the try except block is a good solution, but I do not see any other option. This is something we should review and revise on both python lib and app side.