push primary upload methods to interface. Simplifying Child Classes a lot
This is a big change in previous methods of how data is handled. It simplfies both realtime_database_cli & realtime_database_pyrebase to creating_references / upload, get, and delete data from a database reference, while in the interface class methods are used for handling the data interfaces. This will make maintenance much easier.