update realtime_database_pyrebase.py upload_data_to_reference()
This one I am having a very difficult time with. I was hoping that I could treat it the same as in the _cli..... where I pass it data, serialize the data (with json_dumps(data)), and then upload. However, all of the pushing methods I can find from pyrebase do the json serilization logic within the method. This is causing differences in the methods across both child classes. details in comment.