Refactor file open methods to _cli & _pyrebase modules of Storage package
This is primarily differences and success rate of uploading. I tried to contain the file open and read to the StorageInterface, but the success rate of upload was not very high. If I use python way of parsing byte_data it does poor in the _cli & vise versa for the _pyrebase and trying to access the .NET methods