MAJOR UPDATE on the RealtimeDatabasePyrebase. For data consistency and serilization
This update provides the functionality of passing compas objects all the way to the core functionality. It is a bit odd, because it calls json_dumps and then json.loads, but the logic is that serialization happens in the compas way, and the json.load simply converts the data to an object in a method that will be acceptable for the internal data serialization in the .set() method.